FederatedRealm: Modify generation of initial admin rights.
Initial administrator population is no longer handled by directly
assigning a wilcard permission to the first subject.

A role called 'super administrators' with the wildcard permission is
now used instead with the initial subject being added to this role.

This will allow easier cleanup of access rights in the future.