Initial Population: Plugin provides system user and admin role.
Previously implementing applications had to insert a generic user as
part of their boostrap to do things like create initial workflows.
This is now handled by the plugin.

As part of this bootstrap an all privileges admin role is created for
use by host applications.

The changes FederatedRealm so that in no longer considers
auto_populate. There was always discomfort with this approach on my
side and i've now removed it fully. To set initial admin rights a
database query will need to be run to manually add the initial user
to the admin role as part of initial deployment.