Override Shiro default token strategy.
A we want to see a rich hierachy of exceptions in order to respond to
users when using more then one token type (federated/api at present)
this strategy will ensure we can do this.

If we don't use this strategy as soon as we use more then one token
type all exceptions come out as a single type preventing redirection
to appropriate error messages.