Provides authenticated API to AAF applications.
Implements a HMAC based security layer for our REST API's heavily
influenced by the work of Amazon AWS amongst others.

Currently a web interface to manage ApiSubjects is not provided and
test cases need to be completed. Both of these are left as TODO for
post VHR test environment deployment.