Console: Correct fault with X-FRAME-OPTIONS and console access.
With the introduction of X-FRAME-OPTIONS Deny security filter the
console is no longer able to be shown in a wrapped iFrame.

This commit changes the behaviour to show the console when required
in its own browser tab/window to prevent the X-FRAME-OPTIONS fault.
Further it remaps the show URL to be part of the administration
structure for other administration views.

Closes #2.