Fixed development login attributes to avoid errors while developing attributevalidator
Changed AEPST to the correct format as documented, using the SHA1 hash
of some random data, and then permuting the last character of the hash
to ensure all the values were different.

Fixed a couple of bad attributes for Johnny Moreattr too, as it was
intended that this user would have a mostly complete set of valid
attributes to be used during development, and two of them weren't
correct.